Compute framework coverage diff in artifacts job
